+++ app/src/main/java/com/example/wearme/domain/model/api/Measurements.kt
@@ -3,8 +3,8 @@ package com.example.wearme.domain.model.api
 import com.google.gson.annotations.SerializedName
 
 data class Measurements(
-  @SerializedName("chest") val chest: Int,
-  @SerializedName("waist") val waist: Int,
-  @SerializedName("hips") val hips: Int,
-  @SerializedName("foot") val foot: Int
+    @SerializedName("chest") val chest: Int,
+    @SerializedName("waist") val waist: Int,
+    @SerializedName("hips") val hips: Int,
+    @SerializedName("foot") val foot: Double
 )
